gemiTask/requirements.txt
argparse
google-generativeai
python-dateutil
setuptools
